@@ -15,7 +15,7 @@ namespace ReadMLB.DataLayer.Repositories
         Task<IEnumerable<TEntity>> FindAsync(Expression<Func<TEntity, bool>> predicate);
 
         Task<IEnumerable<TEntity>> FindAsync<TKey>(Expression<Func<TEntity, bool>> predicate, Expression<Func<TEntity, TKey>> orderBy = null, int? take = 50, int? skip = 0);
-
+        Task<IEnumerable<TEntity>> FindAsync<TInc, TKey>(Expression<Func<TEntity, TInc>> include, Expression<Func<TEntity, bool>> predicate, Expression<Func<TEntity, TKey>> orderBy = null, int? take = 50, int? skip = 0);
         Task<TEntity> SingleOrDefaultAsync(Expression<Func<TEntity, bool>> predicate);
 
         ValueTask<EntityEntry<TEntity>> AddAsync(TEntity entity);
