@@ -2,7 +2,10 @@
 using ReadMLB.DataLayer.Repositories;
 using ReadMLB.Entities;
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
+using System.Linq.Expressions;
 using System.Threading.Tasks;
 
 namespace ReadMLB.Services
@@ -10,6 +13,12 @@ namespace ReadMLB.Services
     public interface IBattingService
     {
         Task<long> AddBattingStatAsync(Batting battingStat);
+
+        Task<ICollection<Batting>> GetTeamBattersAsync(short year, short teamId);
+
+        Task<ICollection<Batting>> GetOrganizationBattersAsync(short year, byte organizationId);
+
+        Task<ICollection<Batting>> GetTopLeagueBattersAsync<TKey>(short year, byte league, Expression<Func<Batting, TKey>> category, short top = 50);
 
         Task TruncateBattingStatsAsync();
     }
@@ -23,17 +32,27 @@ namespace ReadMLB.Services
         }
         public async Task<long> AddBattingStatAsync(Batting battingStat)
         {
-            try
-            {
                 var result = await _unitOfWork.BattingStats.AddAsync(battingStat);
                 await _unitOfWork.CompleteAsync();
                 return result.Entity.BattingId;
-            }
-            catch (Exception ex)
-            {
-                return 0;
-            }
-            
+        }
+
+        public async Task<ICollection<Batting>> GetOrganizationBattersAsync(short year, byte organizationId)
+        {
+            var batters = await _unitOfWork.BattingStats.FindAsync(b => b.Team, b => b.Year == year && b.Team.OrganizationId == organizationId, b => b.PlayerId);
+            return batters.ToList();
+        }
+
+        public async Task<ICollection<Batting>> GetTeamBattersAsync(short year, short teamId)
+        {
+            var batters = await _unitOfWork.BattingStats.FindAsync(b => b.Player, b => b.Year == year && b.TeamId == teamId, b => b.PlayerId);
+            return batters.ToList();
+        }
+
+        public async Task<ICollection<Batting>> GetTopLeagueBattersAsync<TKey>(short year, byte league, Expression<Func<Batting,TKey>> category, short top = 50)
+        {
+            var batters = await _unitOfWork.BattingStats.FindAsync<TKey>(b => b.Year == year && b.League == league, category, top);
+            return batters.ToList();
         }
 
         public Task TruncateBattingStatsAsync()
